Crate example_login

Enums

LoginChoice 🔒

Constants

INITIAL_DEVICE_DISPLAY_NAME 🔒

The initial device name when logging in with a device for the first time.

Functions

login_and_sync 🔒

Log in to the given homeserver and sync.

login_with_password 🔒

Login with a username and password.

login_with_sso 🔒

Login with SSO.

main 🔒

A simple program that adapts to the different login methods offered by a Matrix homeserver.

offer_choices_and_login 🔒

Offer the given choices to the user and login with the selected option.

on_room_message 🔒

Handle room messages by logging them.